Player and Game Mode
This implements the Player Character. Camera movement Yaw = Mouse X,
Gamepad Right Thumb Stick X. and Pitch = Mouse Y, Gamepad Right Thumb
Stick Y(Needs Work). I implemented a dummy Game mode but will be added
later on to keep track of the game state and player points, clean up of
level to be able to replay it. It also sets the Player Character as the
default Player Pawn so we can spawn it.